Refactor cut-and-paste code in histogramMsg
This does a refactor of the histogram code to avoid code duplications
with minor variations in the datatype.  The approach I took was to
write a nested procedure (to avoid having to pass arguments because
I'm lazy) that only takes a single 'type' argument which is used to
do the casting.