Update `entry` argument in `addEntry()` to have `in` intent for Chapel 1.20
In Chapel 1.20, `shared` and `owned` types will have `const ref`
intent by default in order to avoid (typically) unnecessary accidental
ownership transfers or reference counting.  In Chapel 1.19, these
types had `in` intent by default which resulted in other challenges
and surprises.

Here, I've inserted an explicit `in` intent for the `entry` argument
in `addEntry()` to preserve the historic behavior and avoid an error
message that otherwise comes up in Chapel 1.20:

  error: value from coercion passed to const ref formal 'x'

The coercion that this is referring to relates to the passing of
`new shared SymEntry()` due to it being a subclass of `GenSymEntry`.

(Aside: My emacs was having trouble with linefeeds in this file and
seemed to auto-identify it as a DOS file for some reason...  I'm
not convinced I haven't made things worse).