st.addEntry returns borrowed entry into table
This commit changes the insertion behavior of GenSymEntry.addEntry().

The first lookup into the table is saved as a reference. A new entry can
be constructed into that slot and we can now return a reference into the
table of the new entry without a second lookup.

As such, addEntry() now returns a borrow to the entry that was just
inserted so that subsequent operations in OperatorMsg don't have to
explicitly look up the entry again.

If using the typed form of addEntry (like most of the code in
OperatorMsg), the returned borrow is of the correct SymEntry type, which
also removes the need to explicitly cast.