Remove need for LLVM for Arkouda
HDF5 no longer requires LLVM to build as of Chapel 1.20.0, so remove
that line from the README.  The extern block in RadixSortLSD does need
LLVM, but it seems that routine could be written in Chapel as shown
here.  So we shouldn't need LLVM anymore.  This permits arkouda to be
built with a homebrew installation of Chapel (single locale only)