Convert arkouda to nilable (one option for making Arkouda Chapel 1.20-ready)
This is a first stab at converting Arkouda over to using nilable
vs. non-nilable classes.  Generally speaking, the approach I took was
to use non-nilable classes for the most part and to rely on throwing
and catching errors to handle failed lookups or conversions.  Note
that I didn't worry much about getting indenting "right" in order to
try and minimize the diff against master for the sake of comparison by
human readers. If we go forward with this PR, we should update the
indentation either before or after merging.

At this point, I'm opening this PR to get feedback on the approach
taken and whether the arkouda keepers would like to see a different
approach or any other minor changes.

Note: There's a new reflection function that can be used to get the
name of the function in which it's called, and this might be nice to
use rather than hard-coding all of the procedure names (or capturing
them in `pn`), but I didn't undertake that here. Just mentioning it
for a separate PR.