Add Building arkouda server to pip install
This commit adds the class to override the setup.py install
command so we can supply one that also builds chapel and
the arkouda server.

The run method now clones, configures and builds chapel as
well as the arkouda server.

This can be extended to support all the setup.py commands
including build, develop etc.