Make Arkouda's `use` statements work if `use` => `private use` by default
Here, I've massaged Arkouda's `use` statements such that they will
work if we take the proposed action of having `use` mean `private use`
by default.  In some cases, I added new `use` statements to a given
module; in others, I made an existing `use` into a `public use`.  I
don't feel expert in understanding the organization of the Arkouda
code, so am not confident that I've made the best choices in all cases,
but this seems to compile and run in a "private use by default" world,
so that's a start.