Add support for src/TAGS for Chapel source
This adds a `make tags` rule to the top-level Arkouda Makefile
which attempts to run $CHPL_HOME/util/chpltags on the src/
directory in order to create a TAGS file for use with emacs
and/or vim.  This makes it much easier, when editing Arkouda
source, to jump to and find the definitions of symbols without
relying on grep or foreknowledge of the code organization.