Add basic travis-ci testing
Run tests/checks.py with a comm=none server build.

This uses the Chapel 1.20 comm=none docker image. The script installs
the dependencies (hdf5, zmq, python3 and python deps.) hdf5 is not
installed to a "normal" location, but is instead installed in a
`hdf5/serial` subdir, so we have to add that to `Makefile.paths` so the
server can find it. The server is built without `--fast` to speed up
backend compilation times and to run with Chapel's checks enabled.
check.py was modified to set an exit code to detect failures and
underscores were removed from numbers since that was only added in
python 3.6 and the travis config is using python 3.5.