Add a mechanism to write the server hostname and port to a file
Add `--serverConnectionInfo` (`ARKOUDA_SERVER_CONNECTION_INFO`) that
will have the server write connection info (hostname:port) out to a
file. When running the server and client on a system with a shared
file system this should make it a lot easier to share hostname/port
information automatically instead of having to manually copy the values
each time.