Override build instead of install; install to py prefix
Previously, the arkouda server was built by overriding the
install process. The arkouda server is now built by overriding
the build_py command which is used in both the supported and
semi-depicated methods of installing through pip.

The arkouda_server executable, upon successful build, is moved
to the bin/ directory of the current python prefix.