Subject: Add ak.sort()
Description:

ENH: This mod adds a new function, ak.sort(), which takes a pdarray
and returns a sorted copy of the array.  The function is nearly
identical to ak.argsort() and uses the same RadixSortLSD to do the
sorting on the Chapel side.

This mod is intended to resolve issue #155 (add ak.sort() to sort
values in a pdarray).