Restore `use IO` statements and make ServerConfig's `public`
This re-instates the `use IO` statements introduced in PR #189
and commented out in PR #196.  Though they are not required
with Chapel 1.20, they are with the master branch of Chapel.

In addition, I made the use statement in ServerConfig public
such that it would be transitively available to modules that
`use ServerConfig` as we make `use` statements private by
default.

Finally, I added one new `use IO` statement to ServerErrorStrings
which was necessary for it to get the string.format function.