Optimize and enable aggregation for ugni
Optimize the DstAggregator and use it for sorting under ugni. The code
used to use a single on-stmt that would allocate remote memory, GET, and
process the data. Here we change to on-stmt to allocate remote memory, a
PUT, and a second on-stmt to process the data. The previous approach led
to too many tasks existing concurrently under ugni, which increased
memory usage. This adds an extra on-stmt in order to have shorter lived
tasks. This also fixes a performance bug that led to yielding every time
instead of only every ~4096 times. These efforts improve performance of
DstAgggregator, which is the workhouse of sorting.

Note that we only use the DstAggregator under ugni for the main data
transposition, not for the steps that collect the metadata because the
overhead to allocate+setup the aggregators outweighs the comm benefits
for relatively small amounts of transfers.

This results in ~2x speedup for rank sorting and a ~1.5x speedup for key
sorting at 256 nodes under ugni. This also results in a ~1.5x speedup
for rank and key sorting at 32 nodes under gn-ibv-large.

Full scaling results shown below:
 - Up to 512 nodes of a Cray-XC (36-core Broadwell, 128GB RAM, Aries)
 - Up to  32 nodes of a Cray-CS (36-core Broadwell, 128GB RAM, FDR IB)

```
chpl test/UnitTestSort.chpl --fast -M src/
./UnitTestSort --mode=testMode.performance -nl $nodes
```

radixSortLSD_ranks:
---

Cray-XC (ugni):

| nodes | before    | after    |
| ----- | --------- | -------- |
|   1   | 755 MB/s  | 830 MB/s |
|   2   | 380 MB/s  | 735 MB/s |
|   4   | 280 MB/s  | 640 MB/s |
|   8   | 245 MB/s  | 575 MB/s |
|  16   | 225 MB/s  | 540 MB/s |
|  32   | 220 MB/s  | 525 MB/s |
|  64   | 215 MB/s  | 510 MB/s |
| 128   | 210 MB/s  | 470 MB/s |
| 256   | 205 MB/s  | 440 MB/s |
| 512   | 195 MB/s  | 365 MB/s |

Cray-CS (gn-ibv-large):

| nodes | before    | after    |
| ----- | --------- | -------- |
|   1   | 350 MB/s  | 830 MB/s |
|   2   | 425 MB/s  | 715 MB/s |
|   4   | 345 MB/s  | 545 MB/s |
|   8   | 310 MB/s  | 480 MB/s |
|  16   | 290 MB/s  | 465 MB/s |
|  32   | 270 MB/s  | 415 MB/s |

radixSortLSD_keys:
---

Cray-XC (ugni):

| nodes | before    | after    |
| ----- | --------- | -------- |
|   1   | 980 MB/s  | 990 MB/s |
|   2   | 630 MB/s  | 800 MB/s |
|   4   | 495 MB/s  | 680 MB/s |
|   8   | 430 MB/s  | 635 MB/s |
|  16   | 400 MB/s  | 595 MB/s |
|  32   | 385 MB/s  | 580 MB/s |
|  64   | 375 MB/s  | 570 MB/s |
| 128   | 360 MB/s  | 550 MB/s |
| 256   | 345 MB/s  | 505 MB/s |
| 512   | 335 MB/s  | 445 MB/s |

Cray-CS (gn-ibv-large):

| nodes | before    | after    |
| ----- | --------- | -------- |
|   1   | 400 MB/s  | 990 MB/s |
|   2   | 430 MB/s  | 715 MB/s |
|   4   | 355 MB/s  | 545 MB/s |
|   8   | 315 MB/s  | 480 MB/s |
|  16   | 295 MB/s  | 465 MB/s |
|  32   | 285 MB/s  | 415 MB/s |