Add `use SysCTypes` where needed
Traditionally, types in SysCTypes have been automatically available to
Chapel code, but as part of our effort to be more precise about
namespaces, we have changed that so that an explicit `use SysCTypes;`
is now required.  This PR adds such a statement to the places in
Arkouda that currently rely on these types and don't already have them
available.