Add a qualified reference to Time.Timer
On the master branch of Chapel, we're getting much better about
reducing the number of symbols that are inadvertently made available
in user code.  This PR updates a reference to `Timer` to be the
fully-qualified `Time.Timer` since the use statement referring to Time
is `use Time only;` (alternatively, one could presumably change it to
just `use Time;` or `use Time only Timer;` but I didn't try either of
those for the sake of expediency).