Avoid type-punning in getDigit for reals
`getDigit` for reals was trying to convert the real to a uint and
then shift that. However, it was converting with casts that resulted
in type-punning which is UB. Instead use memcpy to cast. This is the
right way to convert types, and all C compilers will optimize this
pattern (we have used this idiom and verified the asm is optimal in
our runtime.)