Improve UnitTestSort performance mode
Disable rank sort verification when doing performance testing. It ends
up using fine-grained comm to gather elements of A, which is slow for
the same configurations aggregation is needed in the first place. We
could use aggregation to make the verification faster, but seems like
there could be a smarter way to verify.

Also add some warmup runs to get slightly more consistent timings.