Optimize sorting negative values
Sort negatives directly instead of reordering them after sorting.
Previously, the radix sort code would result in negatives being reversed
and on the right hand side of the array. The old code would then perform
a final permutation to get them in the right order, but this added a lot
of extra comm. Here we use some bit tricks so negatives sort correctly.

For ints, we just invert the signbit so negatives get sorted correctly.
Because of the other getBitWidth optimization, we have to track if we're
sorting the last digit and if there was a signbit to flip (were there
any negatives.) Without the getBitWidth trick, this would be something
like:

```chpl
key ^= (1<<63);
```

For reals, we invert the signbit for positive numbers, and invert all
the bits for negative numbers. Without the getBitWidth trick this would
be something like:

```chpl
if signbit(key) {
  key = ~key;
} else {
  key ^= (1<<63);
}
```