Improve getBitWidth for min(int(?)) values
For values like `min(int(8))`, the old code would say that 9 bits are
needed. It would invert -128, and compute that 8 bits are needed for
128, and then add 1 for the signbit. This was wrong because there's 1
more negative value that can be represented for a given bit range. Just
subract 1 from our inverted value before calling clz to fix this. This
also allows us to get rid of the code to make sure we're using at most
64 bits. That was a workaround for the old code computing that
`min(int(64))` needed 65 bits, but it now correctly computes 64.