Migrate from Travis-CI to Github actions
Travis has lower RAM limits and no swap, which has led to OOMs for the
compiler. Switch over to github actions, which has 7GB of RAM and 7GB of
swap.

I pieced this together with trial and error and using:
 - https://help.github.com/en/actions/automating-your-workflow-with-github-actions
 - https://help.github.com/en/actions/automating-your-workflow-with-github-actions/workflow-syntax-for-github-actions

Personally, I think this is cleaner than what we had before, and it's
using a native github feature instead of something external.