Add a script that runs and optionally graphs benchmarks
This adds a script that runs the benchmarks in the benchmark directory
and optionally uses the Chapel graphing infrastructure to generate
interactive graphs. Example usage, running the benchmarks on 16 nodes of
an XC:

    ./benchmarks/run_benchmarks.py -nl 16

And to generate graphs:

    ./benchmarks/run_benchmarks.py -nl 16 --gen-graphs --platform-name "16 node XC"

The driver starts the server, runs all the benchmarks, and stops the
server. If graph generation is enabled, the time and rate from the
benchmarks is written to ".dat" files using a chapel "computePerfStats"
script. A chapel "genGraphs" script is then used to process the
generated ".dat" files and some committed ".graph" files to produce
interactive graphs that will be hosted online.

This is meant to support nightly/automated performance testing, but I
have found the driver useful for comparing before/after performance for
a change.