Add a mechanism to write the server hostname and port to a file
Add `--serverConnectionInfo` (`ARKOUDA_SERVER_CONNECTION_INFO`) that
will have the server write connection info (`hostname port`) out to a
file. When running the server and client on a system with a shared file
system this should make it easier to share hostname/port information
automatically instead of having to manually copy the values each time.

With this you can do something like:

```
./arkouda_server -nl 16 --serverConnectionInfo ak-server-info
./benchmarks/stream.py $(cat ak-server-info)
```