Add `make check`
Add util.py, that has functionality to start the server, stop the
server, run a client, pick a default number of locales based on how
arkouda was compiled, and several other options. Use this to implement a
script that starts the server, runs check.py, and stops the server. This
script is called by `make check`