Remove throws from SipHash
It's no longer needed now that outlen is a param. This eliminates a
conditional at the callsite, which has a slight performance improvement
(189 -> 198 MB/s) for the following serial run:

    ./SipHashSpeedTest -nl 1 --NINPUTS=1_000_000 --INPUTSIZE=8 --dataParTasksPerLocale=1

See Chapel issue 9380 for more information on error-handling overhead.
Note that error-handling overhead is generally negligible, but the hash
kernel is called millions of times and is generally very fast, so minor
overheads add up. For `--INPUTSIZE=1024` there is no noticeable
performance difference.