Use bulk comm to localize regions (if they live on one locale)
The old version would do a 1-byte GET for each element we needed, but
here we can do bulk comm. This does add an allocation (which we might be
able to optimize out with scratch space), but the benefit of being able
to use bulk comm (and then using c_ptr) far outweighs the allocation
cost.