Switch from workaround to here.hostname
[NOTE: Do not merge until Arkouda switches to Chapel 1.21]

With the upcoming Chapel 1.21 release, the locale type will support a
`here.hostname` query that should make the workaround that Arkouda has
traditionally been using no longer necessary.

As the note above says, this shouldn't be merged until Arkouda is
ready to use Chapel 1.21.  And it would be worth having someone check
that the behavior of this patch is as expected (beyond making sure
things still compiled and ran, I didn't do further checking of it).