Add coargsort correctness and performance tests
Add basic checks to tests/check.py, comprehensive checks with
tests/coargsort.py, and a perf test with benchmarks/coargsort.py.

The correctness tests are pretty extensive and test various bit-values,
positive/negative values, sorting of less-significant arrays, multiple
value sizes/types in the same call, and a few other things.

The benchmark test is modeled after the argsort benchmark. By default it
sorts positive 32-bit values. It will call coargsort on 1, 2, 8, and 16
arrays.

Sorting times are stable (and slow compared to other ops) so use fewer
trials by default.

This is in preparation for upcoming optimizations to coargsort to ensure
we don't break correctness and to track performance improvements.