Allow hdf5 and zeromq paths to be specified by environment variables
This makes it easier to specify hdf5 and zeromq installation
directories. This is primarily motivated by not wanting to have to
rebuild hdf5/zeromq for nightly performance testing, but I also find
this useful for myself while developing. I run on a bunch of systems and
have a couple concurrent arkouda builds for different performance
experiments and setting the env var once is simpler than creating
Makefile.paths manually every time.