Add tunneling support
Use PyZMQs ssh tunneling to support tunneling zeromq connections. Note
that PyZMQ tunneling requires pexpect, so I have added that as a test
dependency for now.

This makes it easy to run client programs on a machine not directly
connected to the server. For instance, I can run a server on a Cray XC
or CS and run clients on my laptop. This also enables running from a
Cray XC elogin (external login) node, which was the original motivation
for this. In nightly performance testing I want to build/launch from an
elogin node since they have more memory and a physical disk, but the
elogin can't talk directly to the compute nodes so we need to proxy
through the login node. elogin nodes are physically close to the login
node so this does not add much latency.

For more info on pyzmq ssh tunneling see:
 - https://pyzmq.readthedocs.io/en/latest/ssh.html
 - https://pyzmq.readthedocs.io/en/latest/api/zmq.ssh.tunnel.html