Add run_client timeout support
Add the ability to timeout the clients and server shutdown for automated
and nightly testing. Because of current incompatibles between Arkouda
master and Chapel master the server is crashing, which means clients
will hang forever in arkouda.connect(). Add basic support for timing out
the clients so our testing doesn't hang forever, and add support for
timing out the server shutdown since that includes a connect call.