Make SegStringSort.chpl compatible with Chapel 1.20-1.22
This seems to be the final change required for me to run
Arkouda against check.py, groupby_test.py, and string.py
on 1.20-1.22.  Like my change in PR #351, it relies on
playing a game that I'm not happy about which could be
much cleaner once we stop worrying about supporting 1.20.
Namely, I use a little code snippet to infer whether the
current compiler is using 0- or 1-based indexing and then
adjusts some values depending on what it finds.

This routine was also slightly disheartening to work on
because one of the final known things that was _not_
converted to 0-based counting in 1.22 is the "part" sent
to keyPart() routines for Sort (see issue
https://github.com/chapel-lang/chapel/issues/15419),
suggesting that this routine may need another tweak for
1.23...