Default to DefaultRectangular (local array) for comm=none
BlockDist has some overheads that are unnecessary for comm=none builds
(block dist has locality checks, loops over the locales array, and other
things that aren't needed for comm=none.) We're looking to optimize this
in the future, but in the meantime this will improve compilation and
execution performance for single node builds.

The compilation benefits are useful for development. This takes a local
build from 9 minutes to 6 minutes.

This has non-trivial improvements on execution performance for comm=none
mostly from optimizing locality and runtime checks out of array
accesses.

Performance on the 24-core server that nightly perf testing runs on:

| benchmark | BlockDist    | DefaultRect  |
| --------- | -----------: | -----------: |
| argsort   |   0.34 GiB/s |   0.38 GiB/s |
| gather    |   6.40 GiB/s |   7.55 GiB/s |
| reduce    | 104.20 GiB/s | 105.30 GiB/s |
| scan      |   5.05 GiB/s |   6.90 GiB/s |
| scatter   |   4.35 GiB/s |  11.45 GiB/s |
| stream    |   14.0 GiB/s |  21.30 GiB/s |