fixed command to build ak py package --roberto Built chapel 1.20.0 and arkouda from scratch on mac Followed guide on README.md Encountered issue when building the arkouda python import:
"pip3 install -e ./arkouda" generated error: ERROR: File "setup.py" not found. Directory cannot be installed in editable mode

I changed the command to "pip3 install -e ." and it fixed it.
Minor change

(the build instructions told you to be in the repo, but the command assumed you were in parent directory)