Require Chapel 1.21/1.22
Chapel 1.21/1.22 have some significant performance improvements that
benefit Arkouda, especially at scale. Additionally, there are a number
of code cleanups that can be made if we can use 1.21/1.22 so we want to
require this instead of just suggesting it.

I've updated the documentation to say 1.22 is required, and added an
error if 1.20 or older is used. The motivation for allowing 1.21 but
documenting 1.22 is to encourage new users to use 1.22, but allow users
with other Chapel codes that aren't ready for 1.22 to build Arkouda
still.