Use ARKOUDA_QUICK_COMPILE for CI testing
The gasnet CI testing has been hitting out-of-memory conditions recently
and takes a while to compile. This should reduce the memory footprint
(because the backend is using O0 instead of O1) and reduce gasnet build
times from ~30 minutes to ~15 minutes. Total gasnet CI testing time is
around 40 minutes still (25 to run tests) but I'm hoping to lower that
with some upcoming PRs.