fixed OS-related bug in get_username(), added method to creat.arkouda directory, updates per integration testing on MacOS #60
Update VERSION

Need to bump version to create new PyPI package.

Re-add mac CI testing

It was removed in 300 because the infrastructure seemed unstable, but
this might have caught portability issues in 406, so re-enable it to see
how stable things are.

I couldn't find an easy way to install a specific version with homebrew,
and so to pin to chapel 1.22.0 I followed this article:
https://itnext.io/how-to-install-an-older-brew-package-add141e58d32

--re-enabled io_test

readded hdf5-tools per merge issue #60

removed writeln