Improve filename extension detection in write1DDistArray
Previously, the code was just splitting on `.` and assuming anything to
the right of that delimiter was an extension, but that's not always
true. In particular in some of our nightly testing we have directories
with `.` in them, so update the check to not consider something an
extension if it has a path separator in it.

For example our nightly testing path to arkouda is something like
`performance-test-perf.chapcs.arkouda/test/studies/arkouda/arkouda`,
where the `.` in `performance-test-perf.chapcs.arkouda` was skewing this
check previously.