Optimize SrcAggregator by caching remote allocations
This is similar to 485. Previously, the SrcAggregator would allocate
remote arrays in the same on-stmt that would process the buffer. This
led to a lot of repeated array creation/destruction and significantly
increased the lifetime of the remote task. By separately allocating the
remote buffers and caching them, this reduces the amount of allocations
and reduces the lifetime of the remote task.

This results in some non-trivial performance improvements, especially on
high core count systems. We see a 7x improvement for 8 GiB gather on 32
nodes of a 128 core HDR IB system.