Fix some bugs in the mink/maxk implementation
The mink/maxk code maintains a heap while accumulating reduction values
and then as an optimization at the end it gets rid of the heap and sorts
to improve how quickly different tasks can merge their mink/maxk data.
However, it's possible for some accumulation to happen after combining,
so we need to handle that case. It doesn't happen often so doesn't need
to be fast.