Enable SrcAggregator under ugni
Previously, the SrcAggregator was not used under ugni because of memory
pressure at scale. Those issues were resolved in 503.

At 256 nodes for the default problem size (3/4 GiB arrays) this improves
the performance for gathers of int/real by ~1.5x, and gathers of bools
by ~25x. Previously, we were just using unorderedCopy which offers good
performance on XC, but aggregation is better now. For the bool case,
even though we were calling unorderedCopy the Aries NIC can't do RDMA on
things that aren't 4-byte aligned, so we had to fall back to a slow path
that used ordered bounce-buffers. Because of this aggregation offers a
much more significant improvement for bools.

There may still be unaligned comm for the last flush of the bool
aggregation buffer, but the common case will be aligned and it's a bulk
transfer anyway so the performance hit isn't very noticeable. For that
reason I just removed some TODOs to align the flush, instead of rounding
the transfer up.

Perf results for gather on 256 nodes using the default problem size (3/4
GiB arrays):

| type  | before      | after       |
| ----- | ----------: | ----------: |
| bool  |   2.1 GiB/s |  55.3 GiB/s |
| int64 | 271.7 GiB/s | 398.8 GiB/s |

Note that the time for bool/int gather is the same now, the bool rate is
just lower since we're transferring 1/8 as much data.

Resolves 318